Update cohort_analysis.sql
This will fix the problem but the code is still not 100% correct as it does not take into account the scenario where from a certain cohort there are no users in month X but there are in X+1 , the rank() function will not work there. 
This is a real edge case but just to be sure i added alternative version that is a bit slower but should work in this edge case.